Taplist queue now handles merges.

+++ src/Command/LbfUpdateTaplistQueueCommand.php
@@ -5,11 +5,14 @@ namespace App\Command;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Doctrine\ORM\EntityManagerInterface;
 use App\Service\Tools;
 use Symfony\Component\Console\Input\ArrayInput;
 use App\Entity\Event\TapListQueue;
+use App\Service\UntappdAPI;
+use App\Service\UntappdAPISerializer;
 
 class LbfUpdateTaplistQueueCommand extends Command
 {
@@ -22,8 +25,10 @@ class LbfUpdateTaplistQueueCommand extends Command
         ;
     }
     
-    public function __construct(EntityManagerInterface $em, Tools $tools)
+    public function __construct(EntityManagerInterface $em, Tools $tools, UntappdAPI $untappdAPI, UntappdAPISerializer $untappdAPISerializer)
     {
+        $this->untappdAPI = $untappdAPI;
+        $this->untappdAPISerializer = $untappdAPISerializer;
         $this->em = $em;
         $this->em->getConnection()->getConfiguration()->setSQLLogger(null);
         $this->tools = $tools;
@@ -42,31 +47,36 @@ class LbfUpdateTaplistQueueCommand extends Command
             } elseif (array_sum($apiKeyPool) < 5) {
                 $output->writeln(sprintf('[%s] API Key pool is too low (%d)', date('H:i:s'), array_sum($apiKeyPool)));
             } else {
-                $beerInfoCommand = $this->getApplication()->find('untappd:get:beer:info');
                 $output->writeln(sprintf('[%s] Adding beer %d', date('H:i:s'), $queue[0]->getUntappdId()));
-                $arguments = array(
-                    'command' => 'untappd:get:beer:info',
-                    'id'    => $queue[0]->getUntappdId(),
-                    '-e'  => 'prod',
-                    '--no-debug'  => true,
-                );
-                
                 $session = $queue[0]->getSession();
                 
-                $beerInfoCommand->run(new ArrayInput($arguments), $output);
-                $beer = $this->em->getRepository('App\Entity\Beer\Beer')->find($queue[0]->getUntappdId());
-                if ($beer) {
-                    if (!$session->getBeers()->contains($beer)) {
-                        $session->addBeer($beer);
-                        $this->em->persist($session);
+                $apiKey = $this->tools->getBestAPIKey($apiKeyPool);
+                
+                if ($apiKey === false) {
+                    $output->writeln(sprintf('[%s] No more API keys available', date('H:i:s')));
+                    return false;
+                }
+                
+                if ($response = $this->untappdAPI->getBeerInfo($queue[0]->getUntappdId(), $apiKey)) {
+                    $output->writeln(sprintf('[%s] Successfully received beer information', date('H:i:s')));
+                    $beerData = $response->body->response->beer;
+                    $beer = $this->untappdAPISerializer->handleBeerObject($beerData);
+                    $output->writeln(sprintf('[%s] Beer %s has been created/updated', date('H:i:s'), $beer->getName()));
+                    if ($beer) {
+                        if (!$session->getBeers()->contains($beer)) {
+                            $session->addBeer($beer);
+                            $this->em->persist($session);
+                        } else {
+                            $output->writeln(sprintf('[%s] Beer was already in taplist', date('H:i:s')));
+                        }
+                        $this->em->remove($queue[0]);
+                        $this->em->flush();
+                        $output->writeln(sprintf('[%s] Beer has been moved from queue to taplist', date('H:i:s')));
                     } else {
-                        $output->writeln(sprintf('[%s] Beer was already in taplist', date('H:i:s')));
+                        $output->writeln(sprintf('[%s] Couldn\'t find the created beer', date('H:i:s')));
                     }
-                    $this->em->remove($queue[0]);
-                    $this->em->flush();
-                    $output->writeln(sprintf('[%s] Beer has been moved from queue to taplist', date('H:i:s')));
                 } else {
-                    $output->writeln(sprintf('[%s] Couldn\'t find the created beer', date('H:i:s')));
+                    $output->writeln(sprintf('[%s] Couldn\'t get beer information', date('H:i:s')));
                 }
             }
         }
